+++ helper/packages.sh
@@ -8,7 +8,7 @@ DEPENDENCY_PACKAGES=(
 
 function installDependencyPackage() {
     for ((i = 0; i < ${#DEPENDENCY_PACKAGES[@]}; i++)); do
-        echo -e "\nInstalling: ${DEPENDENCY_PACKAGES[i]}\n"
+        log "script" "INFO" "Installing dependency package '${DEPENDENCY_PACKAGES[i]}'"
         sudo apt-get install "${DEPENDENCY_PACKAGES[i]}" -y
     done
 }