+++ install.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 
+SCRIPT_DIR="$(cd "$(dirname "$0")" && pwd)"
 SKIP_INTERACTION=false
 
 for arg in "$@"; do
@@ -14,16 +15,16 @@ for arg in "$@"; do
 done
 
 HELPERS=(
-    "packages" "clone" "dotfiles"
+    "logging" "packages" "clone" "dotfiles"
     "nvchad"
 )
 
 for HELPER in "${HELPERS[@]}"; do
-    source $(pwd)/helper/"${HELPER}".sh
+    source "${SCRIPT_DIR}/helper/${HELPER}.sh"
 done
 
 function main() {
-    echo -e "Running: Update & Upgrade\n"
+    log "script" "INFO" "Running: Update & Upgrade"
     #sudo apt update && sudo apt upgrade
     sudo apt update
     installDependencyPackage
@@ -32,7 +33,7 @@ function main() {
     installDotfiles
     utility
 
-    echo -e "Change Shell to ZSH!"
+    log "script" "INFO" "Change Shell to ZSH!"
     sudo chsh -s $(which zsh) "$USER"
 
     if [ "$SKIP_INTERACTION" = true ]; then
@@ -53,7 +54,7 @@ function main() {
         ;;
     esac
 
-    echo -e "\nInstallation Done :)\n"
+    log "script" "INFO" "Installation Done!"
 }
 
 main