+++ CHANGELOG.md
@@ -1,9 +1,24 @@
-## [1.0.1](https://github.com/armandwipangestu/linux-setup/compare/v1.0.0...v1.0.1) (2025-11-17)
+# [1.1.0-rc.2](https://github.com/armandwipangestu/linux-setup/compare/v1.1.0-rc.1...v1.1.0-rc.2) (2025-11-17)
 
 
 ### Bug Fixes
 
+* make abosulte not relative path source file ([ee6b554](https://github.com/armandwipangestu/linux-setup/commit/ee6b55465c7db3fe9a0eb604cfd4eaa1bf6fa6bd))
+* rename function from logging to log ([40f3fd0](https://github.com/armandwipangestu/linux-setup/commit/40f3fd090b58d59c9bde3c0d0b7e73499e3b8aff))
+
+# [1.1.0-rc.1](https://github.com/armandwipangestu/linux-setup/compare/v1.0.0...v1.1.0-rc.1) (2025-11-17)
+
+
+### Bug Fixes
+
+* make script executable ([f0afac3](https://github.com/armandwipangestu/linux-setup/commit/f0afac3272d9be032e9a5facf672aae4dccaa9af))
 * make script executable ([8af7ee9](https://github.com/armandwipangestu/linux-setup/commit/8af7ee9edb7f8e75c577d9294cb138992c8ce094))
+
+
+### Features
+
+* add logging ([1b1c5b2](https://github.com/armandwipangestu/linux-setup/commit/1b1c5b2f78d80cee095c6d0127d5b501b345f4b0))
+* **ci:** add rule staging branch to publish ([ecfb230](https://github.com/armandwipangestu/linux-setup/commit/ecfb2301f3f822cad1448542ccc421f5c1f52a8c))
 
 # 1.0.0 (2025-11-17)